some cleaning & add static to extern functions

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: kdaumont <kdaumont@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 15:12:05 by kdaumont          #+#    #+#             */
-/*   Updated: 2023/11/07 09:06:47 by kdaumont         ###   ########.fr       */
+/*   Updated: 2023/11/09 12:19:37 by kdaumont         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*mem;
 	char	*base;
