+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: kdaumont <kdaumont@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 15:12:29 by kdaumont          #+#    #+#             */
-/*   Updated: 2023/11/07 09:08:17 by kdaumont         ###   ########.fr       */
+/*   Updated: 2023/11/09 12:23:20 by kdaumont         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strnstr(const char *s1, const char *s2, size_t len)
 {
-	int	i;
-	int	j;
-	int	len_s2;
+	size_t	i;
+	size_t	j;
+	size_t	len_s2;
 
 	if (!s2[0])
 		return ((char *)s1);
